Move DummaryModel from tests to mlc_serve
